refactor: move function rot-8 to a separate module
